+++ packages/core/feature-flags/src/index.ts
@@ -148,6 +148,12 @@ export type FeatureFlags = {
    * a full bundling pass is required based on the AssetGraph's bundlingVersion.
    */
   incrementalBundlingVersioning: boolean;
+
+  /**
+   * Allow for the use of `data-atlaspack-isolated` on script tags in HTML to produce
+   * inline scripts that are build as "isolated" bundles.
+   */
+  inlineIsolatedScripts: boolean;
 };
 
 export type ConsistencyCheckFeatureFlagValue =
@@ -189,6 +195,7 @@ export const DEFAULT_FEATURE_FLAGS: FeatureFlags = {
   condbDevFallbackDev: false,
   condbDevFallbackProd: false,
   incrementalBundlingVersioning: process.env.NODE_ENV === 'test',
+  inlineIsolatedScripts: process.env.NODE_ENV === 'test',
   rustDevServer: false,
 };